copy comments from default config into user config
doesn't copy the header still, but that's not too important, and it would require a more complicated rewrite, since the server's "default config" provided by the bukkit api is a Configuration instance -- not a FileConfiguration -- and only FileConfiguration has the getHeader method. not too sure why. in any case, if i did want to fix this in the future, i could instead read the resource from the jar file myself instead of relying on the bukkit api to give me an instance of the default config